+++ Assets/LowPolyTropicalEnvironment_LITE/Prefabs/Chest/WaterChest.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c4cbf31bc38beb045bf9ac4b601858cb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
