+++ Assets/Prefabs/EnvironmentObjects/Geyser/ChestWater.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0b0ebecc1906a7a4281d4f875fe051af
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
